Revert "Refactor: add modal template story"
This reverts commit a9201060a498df7bb394843c2507b89bfebeffa9.
