+++ src/components/Modals/CardDetailModal/style.ts
